#21 いいかも、いまいちをfindに移行
